wal-utils: fix InMemoryClosableChannelTest and restore tests ()
In , tests were refactored to use assertJ. This refactoring,
caused a compilation error due to the fact that
`assertThat(buffer.get())` expected a `byte` not an `int`.

The compilation error, is however, hidden by the fact that the test
runner silently ignores test that fail discovery.

Since the tests run again, we remove the skipTests tag.

+++ community/testing/wal-utils/src/test/java/org/neo4j/kernel/impl/transaction/log/InMemoryClosableChannelTest.java
@@ -44,10 +44,10 @@ class InMemoryClosableChannelTest {
             channel.read(buffer);
             buffer.flip();
 
-            assertThat(buffer.get()).isEqualTo(0);
-            assertThat(buffer.get()).isEqualTo(1);
-            assertThat(buffer.get()).isEqualTo(2);
-            assertThat(buffer.get()).isEqualTo(3);
+            assertThat(buffer.get()).isEqualTo((byte) 0);
+            assertThat(buffer.get()).isEqualTo((byte) 1);
+            assertThat(buffer.get()).isEqualTo((byte) 2);
+            assertThat(buffer.get()).isEqualTo((byte) 3);
 
             assertThatExceptionOfType(ReadPastEndException.class).isThrownBy(() -> channel.read(buffer));
         }
